requirements.txt
transformers
torch
python-dotenv
PyPDF2
pdf2image
pytesseract
huggingface_hub
accelerate
safetensors
fastapi
uvicorn[standard]
python-multipart
pydantic
python-jose[cryptography]
passlib[bcrypt]